Made a new file for the class Game #44
